requirements.txt
scrapy>=2.11
pymongo>=4.6
python-dotenv>=1.0
pydantic>=2.7
scrapy-playwright>=0.0.35